feat: allow conversions to be individually imported/tree-shakable
