+++ conversions/apple/index.js
@@ -1,0 +1,4 @@
+exports.rgb = require('./rgb');
+
+Object.defineProperty(module.exports, 'channels', {value: 3});
+Object.defineProperty(module.exports, 'labels', {value: ['r16', 'g16', 'b16']});
