+++ conversions/apple/rgb.js
@@ -1,0 +1,3 @@
+module.exports = function (apple) {
+	return [(apple[0] / 65535) * 255, (apple[1] / 65535) * 255, (apple[2] / 65535) * 255];
+};
