+++ conversions/gray/cmyk.js
@@ -1,0 +1,3 @@
+module.exports = function (gray) {
+	return [0, 0, 0, gray[0]];
+};
