+++ conversions/gray/hsl.js
@@ -1,0 +1,3 @@
+module.exports = function (args) {
+	return [0, 0, args[0]];
+};
