+++ conversions/gray/hwb.js
@@ -1,0 +1,3 @@
+module.exports = function (gray) {
+	return [0, 100, gray[0]];
+};
