+++ conversions/gray/lab.js
@@ -1,0 +1,3 @@
+module.exports = function (gray) {
+	return [gray[0], 0, 0];
+};
