+++ conversions/gray/rgb.js
@@ -1,0 +1,3 @@
+module.exports = function (args) {
+	return [args[0] / 100 * 255, args[0] / 100 * 255, args[0] / 100 * 255];
+};
