+++ conversions/hex/index.js
@@ -1,0 +1,4 @@
+exports.rgb = require('./rgb');
+
+Object.defineProperty(module.exports, 'channels', {value: 1});
+Object.defineProperty(module.exports, 'labels', {value: ['hex']});
