+++ conversions/hsv/index.js
@@ -1,0 +1,7 @@
+exports.ansi16 = require('./ansi16');
+exports.hcg = require('./hcg');
+exports.hsl = require('./hsl');
+exports.rgb = require('./rgb');
+
+Object.defineProperty(module.exports, 'channels', {value: 3});
+Object.defineProperty(module.exports, 'labels', {value: 'hsv'});
