+++ conversions/index.js
@@ -1,0 +1,21 @@
+/* MIT license */
+
+// NOTE: conversions should only return primitive values (i.e. arrays, or
+//       values that give correct `typeof` results).
+//       do not use box values types (i.e. Number(), String(), etc.)
+
+exports.rgb = require('./rgb');
+exports.hsl = require('./hsl');
+exports.hsv = require('./hsv');
+exports.hwb = require('./hwb');
+exports.cmyk = require('./cmyk');
+exports.xyz = require('./xyz');
+exports.lab = require('./lab');
+exports.lch = require('./lch');
+exports.hex = require('./hex');
+exports.keyword = require('./keyword');
+exports.ansi16 = require('./ansi16');
+exports.ansi256 = require('./ansi256');
+exports.hcg = require('./hcg');
+exports.apple = require('./apple');
+exports.gray = require('./gray');
