+++ conversions/lch/index.js
@@ -1,0 +1,4 @@
+exports.lab = require('./lab');
+
+Object.defineProperty(module.exports, 'channels', {value: 3});
+Object.defineProperty(module.exports, 'labels', {value: 'lch'});
