+++ conversions/rgb/ansi16.js
@@ -1,0 +1,23 @@
+const rgb2hsv = require('./hsv');
+
+module.exports = function (args, saturation = null) {
+	const [r, g, b] = args;
+	let value = saturation === null ? rgb2hsv(args)[2] : saturation; // Hsv -> ansi16 optimization
+
+	value = Math.round(value / 50);
+
+	if (value === 0) {
+		return 30;
+	}
+
+	let ansi = 30
+		+ ((Math.round(b / 255) << 2)
+		| (Math.round(g / 255) << 1)
+		| Math.round(r / 255));
+
+	if (value === 2) {
+		ansi += 60;
+	}
+
+	return ansi;
+};
