+++ conversions/rgb/apple.js
@@ -1,0 +1,3 @@
+module.exports = function (rgb) {
+	return [(rgb[0] / 255) * 65535, (rgb[1] / 255) * 65535, (rgb[2] / 255) * 65535];
+};
