+++ conversions/rgb/xyz.js
@@ -1,0 +1,16 @@
+module.exports = function (rgb) {
+	let r = rgb[0] / 255;
+	let g = rgb[1] / 255;
+	let b = rgb[2] / 255;
+
+	// Assume sRGB
+	r = r > 0.04045 ? (((r + 0.055) / 1.055) ** 2.4) : (r / 12.92);
+	g = g > 0.04045 ? (((g + 0.055) / 1.055) ** 2.4) : (g / 12.92);
+	b = b > 0.04045 ? (((b + 0.055) / 1.055) ** 2.4) : (b / 12.92);
+
+	const x = (r * 0.4124564) + (g * 0.3575761) + (b * 0.1804375);
+	const y = (r * 0.2126729) + (g * 0.7151522) + (b * 0.072175);
+	const z = (r * 0.0193339) + (g * 0.119192) + (b * 0.9503041);
+
+	return [x * 100, y * 100, z * 100];
+};
