+++ conversions/xyz/index.js
@@ -1,0 +1,5 @@
+exports.lab = require('./lab');
+exports.rgb = require('./rgb');
+
+Object.defineProperty(module.exports, 'channels', {value: 3});
+Object.defineProperty(module.exports, 'labels', {value: 'xyz'});
